requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Pillow==10.1.0
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
